+++ examples/deel-hometask-challenge/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @examples/deel-hometask-challenge
 
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @palmares/databases@0.1.6
+  - @palmares/schemas@0.1.6
+  - @palmares/sequelize-engine@0.1.6
+  - @palmares/zod-schema@0.1.7
+
 ## 0.0.9
 
 ### Patch Changes