+++ libs/drizzle-engine/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @palmares/sequelize-engine
 
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @palmares/databases@0.1.6
+
 ## 0.1.5
 
 ### Patch Changes