+++ libs/zod-schema/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @palmares/zod-schema
 
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @palmares/schemas@0.1.6
+
 ## 0.1.6
 
 ### Patch Changes